winningLotto 프로퍼티 변경, rank 역할 위임
