+++ src/main/kotlin/lottoAuto/domain/WinningLotto.kt
@@ -1,9 +1,20 @@
 package lottoAuto.domain
 
 data class WinningLotto(
-    val winningLottoNumbers: List<LottoNumber>
+    val winningLottoNumbers: List<LottoNumber>,
+    val bonusLottoNumber: LottoNumber
 ) {
-    fun countSameNumber(lotto: Lotto): Int {
+    fun rank(lottoList: List<Lotto>): LottoRanks {
+        val ranks = lottoList.map {
+            LottoRank.from(
+                matchCount = this.countSameNumber(it),
+                withBonus = it.withBonusNumber(bonusLottoNumber)
+            )
+        }
+        return LottoRanks(ranks)
+    }
+
+    private fun countSameNumber(lotto: Lotto): Int {
         return winningLottoNumbers.intersect(
             lotto.lottoNumbers.toSet()
         ).size
